moved everything to result.py + agnostic load
