+++ doc/PARAMS_3_1_2_10_128_0.5
@@ -1,0 +1,1 @@
+ (5.010755805952477049e-02+9.067580118933735552e-06j)  (5.106739134761564824e-02-1.966520817849232012e-05j)  (-8.688637543494535931e-02-5.449990787555081956e-04j)  (-1.120356194126177207e-01+1.450898217734462892e-09j)  (-5.529416639941510392e-02+4.363513397314525209e-05j)  (-4.101189200724969508e-02+2.925116205851226498e-05j)  (2.105922030569331443e-02-5.450003838350430882e-04j)  (-1.913305467033644264e-02+1.566872997645513221e-09j)  (1.466524693894776799e-01-1.071802494068758955e-06j)  (-1.196893476802699391e-01-2.384401540993910648e-05j)  (-8.734109411475428353e-02-5.449830018775564562e-04j)  (-1.008936359346689765e-01+8.979766222826816554e-09j)  (4.607157134265745663e-02+4.311793754352944181e-06j)  (-7.431672941164128854e-02-1.696789163992346375e-05j)  (-1.011133980727411952e-02-5.449789831690018309e-04j)  (-7.595881370300901214e-02+1.043384806038494954e-08j)  (-1.027833354170161473e-01-1.803902676268517500e-05j)  (-1.172735682344239055e-01+5.122748695243078027e-05j)  (4.688352541869590662e-02-5.447586559864848630e-04j)  (-1.092586232830999871e-01+1.355968275342479337e-08j)  (-9.150106166173556377e-02-5.327349374971557346e-06j)  (-3.975973047781512781e-02+3.663193521346406610e-05j)  (9.627384154321236287e-02-5.444030635383291647e-04j)  (-1.215831258901930584e-01+2.626945278302020607e-08j)  (1.013834774728199356e-01+1.105016995333504674e-08j)  (-1.211704643178382335e-01+2.816707379945788276e-08j)  (1.429382734195239202e-01+2.797618114439696359e-07j)  (-9.402843755164555159e-03+1.157778633467400032e-06j)  (1.430289846251829589e-01+3.135164588540730948e-08j)  (3.145637971784047415e-02+1.297453366735777018e-07j)  (7.174844267536015230e-02+8.697593336572606053e-06j)  (-1.383689847333405554e-01+3.445534625284538707e-05j)  (-6.430913368376071915e-02+4.165024520175484495e-08j)  (-1.132275229269734673e-01+1.444617061549073184e-09j)  (-6.118861390448816079e-02-5.761072421359379567e-06j)  (-1.145069305583115704e-01-2.280666103933851367e-05j)  (-5.545382373273395615e-02+2.042448792613293822e-08j)  (-2.643461035486140837e-02+7.084127098360845586e-10j)  (-1.308365488100237395e-01-1.745432681340280617e-05j)  (5.743464997865379423e-02-5.426048163523964463e-05j)  (2.082921319815641409e-02+3.268749604086417885e-07j)  (-6.966964402397145262e-02+1.133750806808223257e-08j)  (6.893909363479772470e-03-2.704952934033506514e-05j)  (-1.221282973846943148e-01-8.430594340339774668e-05j)  (2.363272621874153157e-02+1.583625502967053733e-08j)  (1.295023679821501739e-01+5.492722435919482205e-10j)  (-5.453281374703330964e-02+8.338139473706617104e-06j)  (4.985316471645136194e-02+2.210427783214931953e-05j)  (-1.113094173060266423e-01+5.139968054769298691e-07j)  (6.418465256365067650e-02+1.782779125774173194e-08j)  (-6.328244776742428246e-02+9.435258081082231637e-06j)  (-9.541478369270330051e-02+2.486635617091834601e-05j)  (2.510511009525111595e-02+1.116891769860886120e-06j)  (-1.446812446764600413e-01+3.873963208282857586e-08j)  (9.867948029521116105e-02+3.112449154824698756e-07j)  (-1.485978020396539478e-01+7.932854352586710398e-07j)  (5.334532501061718668e-02+2.805253758224040774e-07j)  (-6.899610175125994727e-02+1.160941744563328994e-06j)  (7.055935134830007560e-02+2.466874887277728596e-08j)  (1.386613007971436440e-01+1.020845698177781917e-07j)  (-7.540339085896288762e-02-1.658705928367696543e-05j)  (2.272592851550387857e-02-6.613064075470459025e-05j)  (2.753194349918695188e-02+1.235034558941741300e-05j)  (2.137062953909438234e-02+3.842445748045691522e-05j)  (-8.317530567810225983e-02-1.362563688276578152e-05j)  (1.354635449017608240e-01-3.711227523760039077e-05j)
