+++ doc/dloss_3_1_2_10_128_0.5
@@ -1,0 +1,1 @@
+6.899104118347167969e-01
