+++ doc/gloss_3_1_2_10_128_0.5
@@ -1,0 +1,1 @@
+7.040155529975891113e-01
